cClient.py
You can find this and the server.py at:
https://medium.com/swlh/lets-write-a-chat-app-in-python-f6783a9ac170
This was easy to get working. You need tkinter and python3, both fairly straightforward installations.
Run it with: python3 cClient.py
Then give it IP: 13.56.125.121
And port: 33000
That will connect to my instance where the server is running. I reboot in the pre-dawn hours, give 
it ten minutes or so and then log back in.
No, no saved names, no saved files, no special permissions, no encryption, it is essentially a telnet 
socket built with python.